netd: add support for ipv6 tethering when using non default apn
When tethering is enabled and a non default apn is used, a policy
route table is created per upstream interface. The table number is
the sum of the interface index and  offset from BASE_TABLE_NUMBER.
There is a rule added per upstream interface which is specific for
the tethered traffic which looks up this table.
This patch adds the capability to pass this table name specific
to the upstream interface to the router advertisement daemon. The
route advertisement daemon will then add the rule which looks up
this table for the ipv6 traffic.

Removed the function addV6RtrAdvIface and removeV6RtrAdvIface
and replaced it with configureV6RtrAdv which has no arguments since
the arguments of the former functions were not being used.

CRs-fixed: 654872
Change-Id: I2845558068ba53032fbbc6be618ac64cfed63c6e