netd: fix concurrency error during iptables execution
When multiple userspace processes run iptables commands concurrently,
some of the commands fail with the following error

08:40:59.164 I/ip6tables( 2308): Another app is currently holding the
xtables lock. Perhaps you want to use the -w option?
08:40:59.165 I/ip6tables( 2308): ip6tables terminated by exit(4)

This is because Iptables uses the xtables lock to handle concurrency.
If the -w option is not used, iptables will exit immediately once it
fails to acquire that lock and this can cause the framework to hang.

CRs-Fixed: 722552
Change-Id: I92281abd5d05745899f9aa3b7fceca31082c47aa